gulp decompress: deal with failures
In a machine without lunzip installed, gulp would fail on transform,
becuase it assumes that decompress ran successfully, and that wasn't
the case, only decompress wasn't validating any errors.

This patch changes gulp decompress to validate that the exec'd command
ends successfully, and prints the error message in case it didn't.